+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.5.0 (TBD)
+
+### New Features
+- Added support for new functions in `snowflake.snowpark.functions`:
+  - `array_generate_range`
+  - `sequence`
+- Added the alias `functions.element_at()` for `functions.get()`
+
 ## 1.4.0 (2023-04-24)
 
 ### New Features